@@ -31,12 +31,11 @@
     "GPL-2.0-or-later"
   ],
   "replace": {
-    "jambagecom/fh-debug": "self.version",
-    "typo3-ter/fh-debug": "self.version"
+    "jambagecom/fh-debug": "self.version"
   },
   "require": {
-    "php": "^7.2",
-    "typo3/cms-core": "^8.7 || ^9.5 || ^10.4 || ^11.5"
+    "php": "^8.0",
+    "typo3/cms-core": "^11.5"
   },
   "autoload": {
     "psr-4": {
@@ -48,7 +47,7 @@
       "extension-key": "fh_debug"
     },
     "branch-alias": {
-        "dev-master": "0.9.x-dev"
+        "dev-master": "0.10.x-dev"
     }
   }
 }
