@@ -30,16 +30,16 @@ class ProductionExceptionHandler extends \TYPO3\CMS\Frontend\ContentObject\Excep
     const TYPO3_DIR         = 'typo3';
 
     /**
-    * Handles exceptions thrown during rendering of content objects
-    * The handler can decide whether to re-throw the exception or
-    * return a nice error message for production context.
-    *
-    * @param \Exception $exception
-    * @param AbstractContentObject $contentObject
-    * @param array $contentObjectConfiguration
-    * @return string
-    * @throws \Exception
-    */
+     * Handles exceptions thrown during rendering of content objects
+     * The handler can decide whether to re-throw the exception or
+     * return a nice error message for production context.
+     *
+     * @param \Exception $exception
+     * @param AbstractContentObject $contentObject
+     * @param array $contentObjectConfiguration
+     * @return string
+     * @throws \Exception
+     */
     public function handle(\Exception $exception, AbstractContentObject $contentObject = null, $contentObjectConfiguration = [])
     {
         if (!empty($this->configuration['ignoreCodes.'])) {
