@@ -8,7 +8,7 @@ $EM_CONF[$_EXTKEY] = [
     'title' => 'Debug generator and Log Writer',
     'description' => 'This generates a PHP debug output file. FileWriter into debug',
     'category' => 'misc',
-    'version' => '0.9.4',
+    'version' => '0.10.0',
     'state' => 'stable',
     'clearcacheonload' => 1,
     'author' => 'Franz Holzinger',
@@ -16,8 +16,8 @@ $EM_CONF[$_EXTKEY] = [
     'author_company' => 'jambage.com',
     'constraints' => [
         'depends' => [
-            'php' => '7.0.0-7.4.99',
-            'typo3' => '8.7.0-11.5.99',
+            'php' => '8.0.0-8.1.99',
+            'typo3' => '11.5.0-11.5.99',
         ],
         'conflicts' => [
         ],
